Add AVX512BW implementation that use vpermi2w

+++ avx512-jpeg-zizag/16bit/common.cpp
@@ -1,6 +1,6 @@
 #include <cstdint>
 
-uint8_t zigzag_reference[64] = {
+uint16_t zigzag_reference[64] = {
      0,  1,  5,  6, 14, 15, 27, 28, 
      2,  4,  7, 13, 16, 26, 29, 42, 
      3,  8, 12, 17, 25, 30, 41, 43, 
@@ -11,7 +11,7 @@ uint8_t zigzag_reference[64] = {
     35, 36, 48, 49, 57, 58, 62, 63
 };
 
-uint8_t zigzag_shuffle[64] = {
+uint16_t zigzag_shuffle[64] = {
      0,  1,  8, 16,  9,  2,  3, 10,
     17, 24, 32, 25, 18, 11,  4,  5, 
     12, 19, 26, 33, 40, 48, 41, 34,
